feat: Add nonce to authentication
The nonce is meant to be stored in the client applications session. It is then transferred to the server as part of the redirect url. The server in turn includes it in the ticket. This binds the ticket to the user session.

This prevents an attacker from authenticating a target as themselves. It also makes it impossible to reuse a leaked ticket in a different browser session (if the session is temper resistent).